MISUV-3590:   * add property-based test for ~validMFACreditDescription;   * drop redundant unit tests in DocumentDetailSpec;
